requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
kafka-python==2.0.2
pyspark==3.4.1
cassandra-driver==3.28.0
mysql-connector-python==8.0.28
pandas==2.1.4
# apache-airflow-providers-apache-spark==4.1.5
# apache-airflow-providers-apache-kafka==1.3.1
# apache-airflow-providers-apache-cassandra==2.1.0 